+++ api.go
@@ -18,52 +18,52 @@ import (
 	"github.com/hypermodeinc/modusdb/api/structreflect"
 )
 
-func Create[T any](driver *Driver, object T, dbId ...uint64) (uint64, T, error) {
-	driver.mutex.Lock()
-	defer driver.mutex.Unlock()
-	if len(dbId) > 1 {
+func Create[T any](engine *Engine, object T, nsId ...uint64) (uint64, T, error) {
+	engine.mutex.Lock()
+	defer engine.mutex.Unlock()
+	if len(nsId) > 1 {
 		return 0, object, fmt.Errorf("only one namespace is allowed")
 	}
-	ctx, db, err := getDefaultDB(driver, dbId...)
+	ctx, ns, err := getDefaultNamespace(engine, nsId...)
 	if err != nil {
 		return 0, object, err
 	}
 
-	gid, err := driver.z.nextUID()
+	gid, err := engine.z.nextUID()
 	if err != nil {
 		return 0, object, err
 	}
 
 	dms := make([]*dql.Mutation, 0)
 	sch := &schema.ParsedSchema{}
-	err = generateSetDqlMutationsAndSchema[T](ctx, db, object, gid, &dms, sch)
+	err = generateSetDqlMutationsAndSchema[T](ctx, ns, object, gid, &dms, sch)
 	if err != nil {
 		return 0, object, err
 	}
 
-	err = driver.alterSchemaWithParsed(ctx, sch)
+	err = engine.alterSchemaWithParsed(ctx, sch)
 	if err != nil {
 		return 0, object, err
 	}
 
-	err = applyDqlMutations(ctx, driver, dms)
+	err = applyDqlMutations(ctx, engine, dms)
 	if err != nil {
 		return 0, object, err
 	}
 
-	return getByGid[T](ctx, db, gid)
+	return getByGid[T](ctx, ns, gid)
 }
 
-func Upsert[T any](driver *Driver, object T, dbId ...uint64) (uint64, T, bool, error) {
+func Upsert[T any](engine *Engine, object T, nsId ...uint64) (uint64, T, bool, error) {
 
 	var wasFound bool
-	driver.mutex.Lock()
-	defer driver.mutex.Unlock()
-	if len(dbId) > 1 {
+	engine.mutex.Lock()
+	defer engine.mutex.Unlock()
+	if len(nsId) > 1 {
 		return 0, object, false, fmt.Errorf("only one namespace is allowed")
 	}
 
-	ctx, db, err := getDefaultDB(driver, dbId...)
+	ctx, ns, err := getDefaultNamespace(engine, nsId...)
 	if err != nil {
 		return 0, object, false, err
 	}
@@ -82,18 +82,18 @@ func Upsert[T any](driver *Driver, object T, dbId ...uint64) (uint64, T, bool, e
 
 	dms := make([]*dql.Mutation, 0)
 	sch := &schema.ParsedSchema{}
-	err = generateSetDqlMutationsAndSchema[T](ctx, db, object, gid, &dms, sch)
+	err = generateSetDqlMutationsAndSchema[T](ctx, ns, object, gid, &dms, sch)
 	if err != nil {
 		return 0, object, false, err
 	}
 
-	err = db.driver.alterSchemaWithParsed(ctx, sch)
+	err = ns.engine.alterSchemaWithParsed(ctx, sch)
 	if err != nil {
 		return 0, object, false, err
 	}
 
 	if gid != 0 || cf != nil {
-		gid, err = getExistingObject[T](ctx, db, gid, cf, object)
+		gid, err = getExistingObject[T](ctx, ns, gid, cf, object)
 		if err != nil && err != apiutils.ErrNoObjFound {
 			return 0, object, false, err
 		}
@@ -101,24 +101,24 @@ func Upsert[T any](driver *Driver, object T, dbId ...uint64) (uint64, T, bool, e
 	}
 
 	if gid == 0 {
-		gid, err = driver.z.nextUID()
+		gid, err = engine.z.nextUID()
 		if err != nil {
 			return 0, object, false, err
 		}
 	}
 
 	dms = make([]*dql.Mutation, 0)
-	err = generateSetDqlMutationsAndSchema[T](ctx, db, object, gid, &dms, sch)
+	err = generateSetDqlMutationsAndSchema[T](ctx, ns, object, gid, &dms, sch)
 	if err != nil {
 		return 0, object, false, err
 	}
 
-	err = applyDqlMutations(ctx, driver, dms)
+	err = applyDqlMutations(ctx, engine, dms)
 	if err != nil {
 		return 0, object, false, err
 	}
 
-	gid, object, err = getByGid[T](ctx, db, gid)
+	gid, object, err = getByGid[T](ctx, ns, gid)
 	if err != nil {
 		return 0, object, false, err
 	}
@@ -126,62 +126,62 @@ func Upsert[T any](driver *Driver, object T, dbId ...uint64) (uint64, T, bool, e
 	return gid, object, wasFound, nil
 }
 
-func Get[T any, R UniqueField](driver *Driver, uniqueField R, dbId ...uint64) (uint64, T, error) {
-	driver.mutex.Lock()
-	defer driver.mutex.Unlock()
+func Get[T any, R UniqueField](engine *Engine, uniqueField R, nsId ...uint64) (uint64, T, error) {
+	engine.mutex.Lock()
+	defer engine.mutex.Unlock()
 	var obj T
-	if len(dbId) > 1 {
+	if len(nsId) > 1 {
 		return 0, obj, fmt.Errorf("only one namespace is allowed")
 	}
-	ctx, db, err := getDefaultDB(driver, dbId...)
+	ctx, ns, err := getDefaultNamespace(engine, nsId...)
 	if err != nil {
 		return 0, obj, err
 	}
 	if uid, ok := any(uniqueField).(uint64); ok {
-		return getByGid[T](ctx, db, uid)
+		return getByGid[T](ctx, ns, uid)
 	}
 
 	if cf, ok := any(uniqueField).(ConstrainedField); ok {
-		return getByConstrainedField[T](ctx, db, cf)
+		return getByConstrainedField[T](ctx, ns, cf)
 	}
 
 	return 0, obj, fmt.Errorf("invalid unique field type")
 }
 
-func Query[T any](driver *Driver, queryParams QueryParams, dbId ...uint64) ([]uint64, []T, error) {
-	driver.mutex.Lock()
-	defer driver.mutex.Unlock()
-	if len(dbId) > 1 {
+func Query[T any](engine *Engine, queryParams QueryParams, nsId ...uint64) ([]uint64, []T, error) {
+	engine.mutex.Lock()
+	defer engine.mutex.Unlock()
+	if len(nsId) > 1 {
 		return nil, nil, fmt.Errorf("only one namespace is allowed")
 	}
-	ctx, db, err := getDefaultDB(driver, dbId...)
+	ctx, ns, err := getDefaultNamespace(engine, nsId...)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	return executeQuery[T](ctx, db, queryParams, true)
+	return executeQuery[T](ctx, ns, queryParams, true)
 }
 
-func Delete[T any, R UniqueField](driver *Driver, uniqueField R, dbId ...uint64) (uint64, T, error) {
-	driver.mutex.Lock()
-	defer driver.mutex.Unlock()
+func Delete[T any, R UniqueField](engine *Engine, uniqueField R, nsId ...uint64) (uint64, T, error) {
+	engine.mutex.Lock()
+	defer engine.mutex.Unlock()
 	var zeroObj T
-	if len(dbId) > 1 {
+	if len(nsId) > 1 {
 		return 0, zeroObj, fmt.Errorf("only one namespace is allowed")
 	}
-	ctx, db, err := getDefaultDB(driver, dbId...)
+	ctx, ns, err := getDefaultNamespace(engine, nsId...)
 	if err != nil {
 		return 0, zeroObj, err
 	}
 	if uid, ok := any(uniqueField).(uint64); ok {
-		uid, obj, err := getByGid[T](ctx, db, uid)
+		uid, obj, err := getByGid[T](ctx, ns, uid)
 		if err != nil {
 			return 0, zeroObj, err
 		}
 
-		dms := generateDeleteDqlMutations(db, uid)
+		dms := generateDeleteDqlMutations(ns, uid)
 
-		err = applyDqlMutations(ctx, driver, dms)
+		err = applyDqlMutations(ctx, engine, dms)
 		if err != nil {
 			return 0, zeroObj, err
 		}
@@ -190,14 +190,14 @@ func Delete[T any, R UniqueField](driver *Driver, uniqueField R, dbId ...uint64)
 	}
 
 	if cf, ok := any(uniqueField).(ConstrainedField); ok {
-		uid, obj, err := getByConstrainedField[T](ctx, db, cf)
+		uid, obj, err := getByConstrainedField[T](ctx, ns, cf)
 		if err != nil {
 			return 0, zeroObj, err
 		}
 
-		dms := generateDeleteDqlMutations(db, uid)
+		dms := generateDeleteDqlMutations(ns, uid)
 
-		err = applyDqlMutations(ctx, driver, dms)
+		err = applyDqlMutations(ctx, engine, dms)
 		if err != nil {
 			return 0, zeroObj, err
 		}