+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - fix: alter schema on reverse edge after querying schema
   [#55](https://github.com/hypermodeinc/modusDB/pull/55)
 
-- feat: update interface to driver and db [#56](https://github.com/hypermodeinc/modusDB/pull/56)
+- feat: update interface to engine and namespace
+  [#56](https://github.com/hypermodeinc/modusDB/pull/56)
 
 ## 2025-01-02 - Version 0.1.0